PROTECTING THAT ASS API

+++ Server/src/main/java/is/hi/hbv501g/kosmosinn/Kosmosinn/RestControllers/CommentRestController.java
@@ -12,6 +12,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 import java.util.List;
@@ -43,24 +46,23 @@ public class CommentRestController {
      */
 
     @PostMapping(value = "/addComment", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public Comment addComment(@PathVariable("topicId") long id, @RequestBody Comment comment) {
-        //User currentUser = (User) session.getAttribute("loggedinuser");
-        //if (comment.getUser().getId() == currentUser.getId()) {
-            Topic topic = topicService.findById(id).get();
-            comment.setTopic(topic);
-            comment.setCommentCreated();
-            comment.setCommentEdited();
-            commentService.save(comment);
-        //}
+    public Comment addComment(@PathVariable("topicId") long id, @RequestBody Comment comment, HttpServletRequest request) {
+        Topic topic = topicService.findById(id).get();
+
+        comment.setUser(userService.findByUserame(comment.getUser().getUsername()));
+        comment.setTopic(topic);
+        comment.setCommentCreated();
+        comment.setCommentEdited();
+        commentService.save(comment);
+        
         return comment;
     }
 
     @PatchMapping(value = "/editComment", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public Comment editComment(@Valid @PathVariable("topicId") long topicId, @Valid @PathVariable("id") long id, @RequestBody Comment editedComment) {
-        User currentUser = (User) session.getAttribute("loggedinuser");
+    public Comment editComment(@Valid @PathVariable("topicId") long topicId, @Valid @PathVariable("id") long id, @RequestBody Comment editedComment, HttpServletRequest request) {
         Topic topic = topicService.findById(topicId).get();
         Comment comment = commentService.findById(id).get();
-        if (userService.isAdmin(currentUser) || editedComment.getUser().getId() == currentUser.getId()) {
+        if (isAllowed(comment, request)) {
             if (editedComment.getCommentText() != null) {
                 comment.setCommentText(editedComment.getCommentText());
                 comment.setCommentEdited();
@@ -70,12 +72,32 @@ public class CommentRestController {
     }
 
     @DeleteMapping("/deleteComment/{id}")
-    public void deleteComment(@Valid @PathVariable("topicId") long topicId, @PathVariable("id") long id) {
+    public void deleteComment(@Valid @PathVariable("topicId") long topicId, @PathVariable("id") long id, HttpServletRequest request) {
         //User currentUser = (User) session.getAttribute("loggedinuser");
         Topic topic = topicService.findById(topicId).get();
         Comment comment = commentService.findById(id).get();
-        //if (userService.isAdmin(currentUser) || comment.getUser().getId() == currentUser.getId()) {
+        if (isAllowed(comment, request)) {
             commentService.delete(comment);
-        //}
+        }
+    }
+
+    private Boolean isAllowed(Comment comment, HttpServletRequest request) {
+        User currentUser = userService.currentUser(request);
+
+        if (userService.isAdmin(currentUser) || commentService.findAllByUserId(currentUser.getId()).contains(comment)) {
+            return true;
+        }
+
+        return false;
+    }
+
+    private Boolean isAllowed(HttpServletRequest request) {
+        User currentUser = userService.currentUser(request);
+
+        if (!currentUser.getRole().isEmpty()) {
+            return true;
+        }
+
+        return false;
     }
 }
